+++ ESLSCDLL/linux/Board_ll.c
@@ -488,7 +488,7 @@ void initPerformanceCounter()
 	return;
 }
 
-es_status_codes SaveMeasurementDataToFileBIN(const char* path, char* filename)
+es_status_codes SaveMeasurementDataToFileBIN(const char* filename)
 {
 	return es_no_error;
 }